Move zonal summary operations to zonal.summary package.
